initial refactor splitting changeloglistene to its own class
